+++ src/Infrastructure/ContainerBuilder.php
@@ -15,12 +15,17 @@ use Sstalle\php7cc\Helper\RegExp\RegExpParser;
 use Sstalle\php7cc\JsonResultPrinter;
 use Sstalle\php7cc\Lexer\ExtendedLexer;
 use Sstalle\php7cc\NodeAnalyzer\FunctionAnalyzer;
+use Sstalle\php7cc\NodeAnalyzer\Reflection\FunctionLike\FunctionCalleeReflector;
+use Sstalle\php7cc\NodeAnalyzer\Reflection\FunctionLike\FunctionLikeCalleeReflector;
+use Sstalle\php7cc\NodeAnalyzer\Reflection\FunctionLike\MethodCalleeReflector;
 use Sstalle\php7cc\NodeStatementsRemover;
 use Sstalle\php7cc\NodeTraverser\Traverser;
 use Sstalle\php7cc\NodeVisitor\Resolver;
 use Sstalle\php7cc\PathChecker;
 use Sstalle\php7cc\PathCheckExecutor;
 use Sstalle\php7cc\PathTraversableFactory;
+use Sstalle\php7cc\Reflection\Internal\Reflector\ClassReflector;
+use Sstalle\php7cc\Reflection\Internal\Reflector\FunctionReflector;
 use Sstalle\php7cc\ResultPrinterInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use PhpParser\PrettyPrinter\Standard as StandardPrettyPrinter;
@@ -184,8 +189,8 @@ class ContainerBuilder
 
             return new Resolver($visitorInstances);
         };
-        $container['nodeAnalyzer.functionAnalyzer'] = function () {
-            return new FunctionAnalyzer();
+        $container['nodeAnalyzer.functionAnalyzer'] = function ($c) {
+            return new FunctionAnalyzer($c['nodeAnalyzer.reflection.functionLikeCalleeReflector']);
         };
         $container['codePreprocessor.shebangRemover'] = function () {
             return new ShebangRemover();
@@ -245,6 +250,24 @@ class ContainerBuilder
         };
         $container['regExpParser'] = function () {
             return new RegExpParser();
+        };
+        $container['reflection.functionReflector'] = function () {
+            return new FunctionReflector();
+        };
+        $container['reflection.classReflector'] = function () {
+            return new ClassReflector();
+        };
+        $container['nodeAnalyzer.reflection.functionCalleeReflector'] = function ($c) {
+            return new FunctionCalleeReflector($c['reflection.functionReflector']);
+        };
+        $container['nodeAnalyzer.reflection.methodCalleeReflector'] = function ($c) {
+            return new MethodCalleeReflector($c['reflection.classReflector']);
+        };
+        $container['nodeAnalyzer.reflection.functionLikeCalleeReflector'] = function ($c) {
+            return new FunctionLikeCalleeReflector(array(
+                $c['nodeAnalyzer.reflection.functionCalleeReflector'],
+                $c['nodeAnalyzer.reflection.methodCalleeReflector'],
+            ));
         };
 
         return $container;