+++ src/NodeAnalyzer/Reflection/Exception/UnsupportedNodeTypeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sstalle\php7cc\NodeAnalyzer\Reflection\Exception;
+
+class UnsupportedNodeTypeException extends \RuntimeException
+{
+}