+++ src/NodeVisitor/FuncGetArgsVisitor.php
@@ -3,6 +3,7 @@
 namespace Sstalle\php7cc\NodeVisitor;
 
 use PhpParser\Node;
+use Sstalle\php7cc\Analysis\Scope\FunctionLikeScope;
 use Sstalle\php7cc\CompatibilityViolation\Message;
 use Sstalle\php7cc\NodeAnalyzer\FunctionAnalyzer;
 
@@ -13,7 +14,7 @@ class FuncGetArgsVisitor extends AbstractVisitor
     /**
      * @var string[]
      */
-    protected $possiblyArgumentModifyingClasses = array(
+    protected $argumentModifyingClasses = array(
         'PhpParser\\Node\\Stmt\\Foreach_',
         'PhpParser\\Node\\Stmt\\Global_',
         'PhpParser\\Node\\Stmt\\Unset_',
@@ -21,7 +22,7 @@ class FuncGetArgsVisitor extends AbstractVisitor
         'PhpParser\\Node\\Expr\\AssignOp',
         'PhpParser\\Node\\Expr\\AssignRef',
         'PhpParser\\Node\\Expr\\FuncCall',
-        'PhpParser\\Node\\Expr\\List',
+        'PhpParser\\Node\\Expr\\List_',
         'PhpParser\\Node\\Expr\\MethodCall',
         'PhpParser\\Node\\Expr\\PostDec',
         'PhpParser\\Node\\Expr\\PostInc',
@@ -36,12 +37,9 @@ class FuncGetArgsVisitor extends AbstractVisitor
     protected $functionAnalyzer;
 
     /**
-     * If current function's arguments could have been modified, value on top of the stack is true.
-     * Otherwise false.
-     *
-     * @var \SplStack
+     * @var FunctionLikeScope[]|\SplStack
      */
-    protected $argumentModificationStack;
+    protected $scopes;
 
     /**
      * @param FunctionAnalyzer $functionAnalyzer
@@ -56,7 +54,7 @@ class FuncGetArgsVisitor extends AbstractVisitor
      */
     public function beforeTraverse(array $nodes)
     {
-        $this->argumentModificationStack = new \SplStack();
+        $this->scopes = new \SplStack();
     }
 
     /**
@@ -64,22 +62,29 @@ class FuncGetArgsVisitor extends AbstractVisitor
      */
     public function enterNode(Node $node)
     {
-        $isCurrentNodeFunctionLike = $node instanceof Node\FunctionLike;
-        if ($isCurrentNodeFunctionLike || $this->argumentModificationStack->isEmpty()
-            || !$this->argumentModificationStack->top()
-            || !$this->functionAnalyzer->isFunctionCallByStaticName($node, array_flip(array('func_get_arg', 'func_get_args')))
-        ) {
-            $isCurrentNodeFunctionLike && $this->argumentModificationStack->push(false);
+        if ($node instanceof Node\FunctionLike) {
+            $parameterNames = $this->extractParametersNames($node->getParams());
+            $this->scopes->push(new FunctionLikeScope($parameterNames));
 
             return;
         }
 
-        /** @var Node\Expr\FuncCall $node */
-        $functionName = $node->name->toString();
-        $this->addContextMessage(
-            sprintf('Function argument(s) returned by "%s" might have been modified', $functionName),
-            $node
-        );
+        $scope = $this->getCurrentScope();
+        if (!$scope
+            || !$this->functionAnalyzer->isFunctionCallByStaticName($node, array_flip(array('func_get_arg', 'func_get_args')))
+        ) {
+            return;
+        }
+
+        $parameterNames = $this->getParameterNamesReturnedBy($node);
+        if ($this->currentFunctionHasModifiedArguments($parameterNames)) {
+            /** @var Node\Expr\FuncCall $node */
+            $functionName = $node->name->toString();
+            $this->addContextMessage(
+                sprintf('Function argument(s) returned by "%s" might have been modified', $functionName),
+                $node
+            );
+        }
     }
 
     /**
@@ -87,23 +92,207 @@ class FuncGetArgsVisitor extends AbstractVisitor
      */
     public function leaveNode(Node $node)
     {
-        if ($this->argumentModificationStack->isEmpty()) {
+        $scope = $this->getCurrentScope();
+        if (!$scope) {
             return;
         }
 
         if ($node instanceof Node\FunctionLike) {
-            $this->argumentModificationStack->pop();
+            $this->scopes->pop();
 
             return;
         }
 
-        foreach ($this->possiblyArgumentModifyingClasses as $class) {
+        foreach ($this->argumentModifyingClasses as $class) {
             if ($node instanceof $class) {
-                $this->argumentModificationStack->pop();
-                $this->argumentModificationStack->push(true);
+                $modifiedVariableNames = $this->getModifiedVariableNames($node);
+                $scope->addModifiedVariables($modifiedVariableNames);
 
                 return;
             }
         }
+    }
+
+    /**
+     * @param Node $node
+     *
+     * @return string[]
+     */
+    private function getModifiedVariableNames(Node $node)
+    {
+        switch (true) {
+            case $node instanceof Node\Stmt\Foreach_:
+                return $this->extractModifiedVariableNames(array($node->keyVar, $node->valueVar));
+
+            case $node instanceof Node\Stmt\Unset_:
+                $modifiedNames = $this->extractModifiedVariableNames($node->vars);
+                $scope = $this->getCurrentScope();
+                foreach ($modifiedNames as $name) {
+                    $scope->unsetVariable($name);
+                }
+
+                return $modifiedNames;
+
+            case $node instanceof Node\Stmt\Global_:
+            case $node instanceof Node\Expr\List_:
+                return $this->extractModifiedVariableNames($node->vars);
+
+            case $node instanceof Node\Expr\Assign:
+            case $node instanceof Node\Expr\AssignOp:
+            case $node instanceof Node\Expr\PostDec:
+            case $node instanceof Node\Expr\PostInc:
+            case $node instanceof Node\Expr\PreDec:
+            case $node instanceof Node\Expr\PreInc:
+                return $this->extractModifiedVariableNames(array($node->var));
+
+            case $node instanceof Node\Expr\FuncCall:
+            case $node instanceof Node\Expr\StaticCall:
+            case $node instanceof Node\Expr\MethodCall:
+                $byReferenceArguments = $this->functionAnalyzer->getByReferenceCallArguments($node);
+
+                return $this->extractModifiedVariableNames($byReferenceArguments);
+
+            case $node instanceof Node\Expr\AssignRef:
+                return $this->addVariableReference($node);
+
+            default:
+                throw new \InvalidArgumentException(sprintf('Unknown node type %s', get_class($node)));
+        }
+    }
+
+    /**
+     * Returns current function's parameter names that could be returned
+     * from a func_get_args/func_get_arg call.
+     *
+     * @param Node\Expr\FuncCall $getArgsCall
+     *
+     * @return string[]
+     */
+    private function getParameterNamesReturnedBy(Node\Expr\FuncCall $getArgsCall)
+    {
+        $scope = $this->getCurrentScope();
+        if (!$scope) {
+            return array();
+        }
+
+        $allParameterNames = $scope->getParameterNames();
+        $argumentNumber = isset($getArgsCall->args[0]) ? $getArgsCall->args[0] : null;
+        if ($argumentNumber
+            && $this->functionAnalyzer->isFunctionCallByStaticName($getArgsCall, 'func_get_arg')
+            && $argumentNumber->value
+            && $argumentNumber->value instanceof Node\Scalar\LNumber
+        ) {
+            $argumentNumber = $argumentNumber->value->value;
+
+            return isset($allParameterNames[$argumentNumber])
+                ? array($allParameterNames[$argumentNumber])
+                : array();
+        }
+
+        return $allParameterNames;
+    }
+
+    /**
+     * @param Node\Expr\AssignRef $node
+     *
+     * @return string[]
+     */
+    private function addVariableReference(Node\Expr\AssignRef $node)
+    {
+        $scope = $this->getCurrentScope();
+
+        /**
+         * If the reference is assigned to an object property, the warning is not relevant,
+         * because the user probably expects the assigned property to be modified.
+         */
+        $isAssignedToVariable = $node->var instanceof Node\Expr\Variable;
+        if (!$scope || !$isAssignedToVariable) {
+            return array();
+        }
+
+        $fromVariableName = $this->convertVariableNameToString($node->expr);
+        $toVariableName = $this->convertVariableNameToString($node->var);
+
+        $scope = $this->getCurrentScope();
+        $scope->createReference($toVariableName, $fromVariableName);
+
+        return array();
+    }
+
+    /**
+     * @param string[] $checkedParameterNames
+     *
+     * @return bool
+     */
+    private function currentFunctionHasModifiedArguments(array $checkedParameterNames)
+    {
+        $scope = $this->getCurrentScope();
+
+        return $scope && $scope->isAnyVariableModified($checkedParameterNames);
+    }
+
+    /**
+     * @param Node\Param[] $parameters
+     *
+     * @return string[]
+     */
+    private function extractParametersNames(array $parameters)
+    {
+        return array_map(function (Node\Param $parameter) {
+            return $parameter->name;
+        }, $parameters);
+    }
+
+    /**
+     * @param Node[] $nodes
+     *
+     * @return string[]
+     */
+    private function extractModifiedVariableNames(array $nodes)
+    {
+        $modifiedNames = array();
+        foreach ($nodes as $node) {
+            if (!$node) {
+                continue;
+            }
+
+            $variableNameNode = $node;
+            if ($node instanceof Node\Arg) {
+                $variableNameNode = $node->value;
+            }
+
+            if ($node instanceof Node\Expr\ArrayDimFetch) {
+                $variableNameNode = $node->var;
+            }
+
+            if (!($variableNameNode instanceof Node\Expr\Variable)) {
+                continue;
+            }
+
+            $variableName = $this->convertVariableNameToString($variableNameNode);
+            $modifiedNames[] = $variableName;
+        }
+
+        return array_unique($modifiedNames);
+    }
+
+    /**
+     * @param Node $variableNode
+     *
+     * @return string
+     */
+    private function convertVariableNameToString(Node $variableNode)
+    {
+        return $variableNode instanceof Node\Expr\Variable && is_string($variableNode->name)
+            ? $variableNode->name
+            : FunctionLikeScope::VARIABLE_VARIABLE_NAME;
+    }
+
+    /**
+     * @return FunctionLikeScope|null
+     */
+    private function getCurrentScope()
+    {
+        return $this->scopes->isEmpty() ? null : $this->scopes->top();
     }
 }