+++ src/Reflection/ReflectionMethodInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sstalle\php7cc\Reflection;
+
+interface ReflectionMethodInterface extends ReflectionFunctionAbstractInterface
+{
+}