+++ src/Reflection/Reflector/ClassReflectorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Sstalle\php7cc\Reflection\Reflector;
+
+use Sstalle\php7cc\Reflection\ReflectionClassInterface;
+
+interface ClassReflectorInterface extends ReflectorInterface
+{
+    /**
+     * @param string $identifer
+     *
+     * @return ReflectionClassInterface
+     */
+    public function reflect($identifer);
+}